Fusion Calculator Component
Finishing up displaying the data for the Fusion Calculator component. Decided to add in another Data class called PageTracker that keeps track of the current page of data for the tables within the Fusion Calculator component. While it is still not a pretty solution, it is better than what I previously had. I still need to work on splitting up some functions so that there is the least amount of intertwined code between the data class and the component class.